Improve docs for `filtering_predicate`.
