+++ src/gpsea/analysis/predicate/genotype/_api.py
@@ -1,7 +1,7 @@
 import abc
 import typing
 
-from gpsea.model import Patient, Variant
+from gpsea.model import Variant
 from .._api import PolyPredicate, Categorization, PatientCategory
 
 
@@ -10,85 +10,7 @@ class GenotypePolyPredicate(PolyPredicate[Categorization], metaclass=abc.ABCMeta
     `GenotypePolyPredicate` is a base class for all :class:`PolyPredicate`
     that test the genotype axis.
     """
-
-    @staticmethod
-    def filtering_predicate(
-        predicate: "GenotypePolyPredicate",
-        targets: typing.Collection[Categorization],
-    ) -> "GenotypePolyPredicate":
-        """
-        """
-        return FilteringGenotypePolyPredicate.create(
-            predicate=predicate,
-            targets=targets,
-        )
-
-
-class FilteringGenotypePolyPredicate(GenotypePolyPredicate):
-    # NOT PART OF THE PUBLIC API
-
-    @staticmethod
-    def create(
-        predicate: "GenotypePolyPredicate",
-        targets: typing.Collection[Categorization],
-    ) -> "FilteringGenotypePolyPredicate":
-        # At least 2 target categorizations must be provided
-        if len(targets) <= 1:
-            raise ValueError(f'At least 2 target categorizations must be provided but got {len(targets)}')
-
-        good_boys = tuple(isinstance(cat, Categorization) for cat in targets)
-        if not all(good_boys):
-            offenders = ', '.join(
-                str(i)
-                for i, is_instance
-                in enumerate(good_boys) if not is_instance
-            )
-            raise ValueError(f'The targets at following indices are not categorizations: [{offenders}]')
-
-        # All `allowed` categorizations must in fact be present in the `base` predicate.
-        cats_are_in_fact_present = tuple(cat in predicate.get_categorizations() for cat in targets)
-        if not all(cats_are_in_fact_present):
-            missing = ', '.join(
-                c.category.name
-                for c, is_present
-                in zip(targets, cats_are_in_fact_present) if not is_present
-            )
-            raise ValueError(f'Some from the categories are not present: {missing}')
-        
-        if len(targets) == predicate.n_categorizations():
-            raise ValueError(
-                f'It makes no sense to subset the a predicate with {predicate.n_categorizations()} categorizations '
-                f'with the same number ({len(targets)}) of targets'
-            )
-
-        return FilteringGenotypePolyPredicate(
-            predicate=predicate,
-            allowed=targets,
-        )
-
-    def __init__(
-        self,
-        predicate: "GenotypePolyPredicate",
-        allowed: typing.Iterable[Categorization],
-    ):
-        self._predicate = predicate
-        self._allowed = tuple(allowed)
-    
-    def get_categorizations(self) -> typing.Sequence[Categorization]:
-        return self._allowed
-
-    def get_question_base(self) -> str:
-        return self._predicate.get_question_base()
-
-    def test(self, patient: Patient) -> typing.Optional[Categorization]:
-        cat = self._predicate.test(patient)
-        if cat in self._allowed:
-            return cat
-        else:
-            return None
-
-    def __repr__(self):
-        return f"FilteringGenotypePolyPredicate(predicate={self._predicate}, allowed={self._allowed})"
+    pass
 
 
 class RecessiveGroupingPredicate(GenotypePolyPredicate, metaclass=abc.ABCMeta):
